pyproject.toml
[project]
name = "urbabel"
authors = [{name = "PTA"}]
version = "0.0.2"
description = "a Python version of Library of Babel with limited multi-language input support"
readme = "README.md"
license = "AGPL-3.0-or-later"
license-files = ["LICENSE"]
classifiers = [
	"Development Status :: 4 - Beta",
	"Operating System :: OS Independent",
	"Environment :: Console",
	"Programming Language :: Python :: 3 :: Only",
	"Programming Language :: Python :: Implementation :: CPython",
	"License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
	"Intended Audience :: Science/Research",
	"Topic :: Scientific/Engineering :: Mathematics",
]
requires-python = ">= 3.14"
dependencies = [
	"gmpy2",
	"pyicu",
]

[project.optional-dependencies]
dev = [
	"coverage",
	"ruff",
	"zuban",
	"nuitka[all]",
]
png = ["pillow"]
pdf = ["fpdf2"]
tui = ["textual"]
webui = ["nicegui"]

[project.urls]
Repository = "https://github.com/phineas-pta/library-babel"

[tool.coverage.run]
omit = ["urbabel/*i.py"]

[tool.ruff.format]
indent-style = "tab"

[tool.ruff.lint]
allowed-confusables = ["×", "’"]
